requirements.txt
playwright>=1.40
yt-dlp>=2025.0.0
pyyaml>=6.0
requests>=2.31
static-ffmpeg>=2.5
streamlit>=1.35
pandas>=2.0
fastapi>=0.110
uvicorn[standard]>=0.27
librosa>=0.11
soundfile>=0.13
scipy>=1.14
